src/requirements.txt
apache-airflow-providers-postgres==5.8.0
pandas>=2.3.3
polars>=1.35.2
pytest>=9.0.0